+++ src/main/java/com/hororok/monta/repository/PaletteRepository.java
@@ -3,7 +3,7 @@ package com.hororok.monta.repository;
 import com.hororok.monta.entity.Palette;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaletteRepository extends JpaRepository<Palette, Long> {
+public interface PaletteRepository extends JpaRepository<Palette, Integer> {
 
 
 }
